Fixed cart functionality
