+++ src/scripts/cart/cart-storage.js
@@ -1,5 +1,4 @@
 // Handle cart storage operations in localStorage
-// This is a separate file because it's used in multiple files
 export function loadCart() {
   const savedCart = localStorage.getItem("cart");
   return savedCart ? JSON.parse(savedCart) : [];
